refactor to use chi
