+++ go.mod
@@ -4,6 +4,7 @@ go 1.13
 
 require (
 	github.com/gernest/front v0.0.0-20181129160812-ed80ca338b88
+	github.com/go-chi/chi v4.0.2+incompatible // indirect
 	github.com/google/uuid v1.1.1
 	github.com/gorilla/mux v1.7.3
 	github.com/shurcooL/sanitized_anchor_name v1.0.0 // indirect
