+++ postmortem.go
@@ -20,6 +20,20 @@ type Postmortem struct {
 	Description string
 }
 
+var (
+	// Categories is a whitelist of valid categories that a postmortem can have.
+	Categories = [...]string{
+		"automation",
+		"cascading-failure",
+		"cloud",
+		"config-change",
+		"postmortem",
+		"security",
+		"time",
+		"undescriptive",
+	}
+)
+
 // Parse turns an io stream into a Postmortem type.
 func Parse(f io.Reader) (*Postmortem, error) {
 	p := &Postmortem{}
@@ -67,4 +81,61 @@ func Parse(f io.Reader) (*Postmortem, error) {
 	p.Description = body
 
 	return p, nil
+}
+
+// GenerateJSON outputs all content in JSON for parsing by our website.
+func GenerateJSON(d string) error {
+	baseDir := "./output"
+
+	err := os.MkdirAll(baseDir, os.ModePerm)
+	if err != nil {
+		return err
+	}
+
+	fp := filepath.Join(baseDir, "categories.json")
+
+	j, err := json.Marshal(postmortems.Categories)
+	if err != nil {
+		return err
+	}
+
+	err = ioutil.WriteFile(fp, j, 0644)
+	if err != nil {
+		return err
+	}
+
+	return filepath.Walk(d, func(path string, info os.FileInfo, err error) error {
+		// Failed to open path
+		if err != nil {
+			return err
+		}
+
+		if !info.IsDir() {
+			f, err := os.Open(path)
+			if err != nil {
+				return err
+			}
+
+			fName := filepath.Base(path)
+			extName := filepath.Ext(path)
+			id := fName[:len(fName)-len(extName)]
+
+			p, err := Parse(f)
+			if err != nil {
+				return err
+			}
+
+			fp := filepath.Join(baseDir, fmt.Sprintf("%s.json", id))
+			j, err := json.Marshal(p)
+			if err != nil {
+				return err
+			}
+			err = ioutil.WriteFile(fp, j, 0644)
+			if err != nil {
+				return err
+			}
+		}
+
+		return nil
+	})
 }
